# [mini] Fix check for FastAllocateString that used old name