The check that was supposed to take managed allocation code path for strings was broken since e58097e957af9939c94ee46ecd78e38b16a1f989. It went unnoticed for over 4 years.