@filipnavara 2019-10-08
The test failures are related. Looking into them.

```
1) Constructor4_LargeString (MonoTests.System.StringTest.Constructor4_LargeString)
   System.OutOfMemoryException : Insufficient memory to continue the execution of the program.
  at (wrapper alloc) System.Object.AllocString(intptr,int)
  at System.String.Ctor (System.Char c, System.Int32 count) [0x0001d] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/external/corefx/src/Common/src/CoreLib/System/String.cs:302 
  at System.String.CreateString (System.Char c, System.Int32 count) [0x00000] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/ReferenceSources/String.cs:373 
  at (wrapper managed-to-managed) System.String..ctor(char,int)
  at MonoTests.System.StringTest.Constructor4_LargeString () [0x00005] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/Test/System/StringTest.cs:126 
  at (wrapper managed-to-native) System.Reflection.RuntimeMethodInfo.InternalInvoke(System.Reflection.RuntimeMethodInfo,object,object[],System.Exception&)
  at System.Reflection.RuntimeMethodInfo.Invoke (System.Object obj, System.Reflection.BindingFlags invokeAttr, System.Reflection.Binder binder, System.Object[] parameters, System.Globalization.CultureInfo culture) [0x0006a] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/System.Reflection/RuntimeMethodInfo.cs:395 


2) PadLeft_LargeString (MonoTests.System.StringTest.PadLeft_LargeString)
   System.OutOfMemoryException : Insufficient memory to continue the execution of the program.
  at (wrapper alloc) System.Object.AllocString(intptr,int)
  at System.String.PadLeft (System.Int32 totalWidth, System.Char paddingChar) [0x00025] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/external/corefx/src/Common/src/CoreLib/System/String.Manipulation.cs:827 
  at MonoTests.System.StringTest.PadLeft_LargeString () [0x00005] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/Test/System/StringTest.cs:3011 
  at (wrapper managed-to-native) System.Reflection.RuntimeMethodInfo.InternalInvoke(System.Reflection.RuntimeMethodInfo,object,object[],System.Exception&)
  at System.Reflection.RuntimeMethodInfo.Invoke (System.Object obj, System.Reflection.BindingFlags invokeAttr, System.Reflection.Binder binder, System.Object[] parameters, System.Globalization.CultureInfo culture) [0x0006a] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/System.Reflection/RuntimeMethodInfo.cs:395 


3) PadRight_LargeString (MonoTests.System.StringTest.PadRight_LargeString)
   System.OutOfMemoryException : Insufficient memory to continue the execution of the program.
  at (wrapper alloc) System.Object.AllocString(intptr,int)
  at System.String.PadRight (System.Int32 totalWidth, System.Char paddingChar) [0x00025] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/external/corefx/src/Common/src/CoreLib/System/String.Manipulation.cs:853 
  at MonoTests.System.StringTest.PadRight_LargeString () [0x00005] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/Test/System/StringTest.cs:3063 
  at (wrapper managed-to-native) System.Reflection.RuntimeMethodInfo.InternalInvoke(System.Reflection.RuntimeMethodInfo,object,object[],System.Exception&)
  at System.Reflection.RuntimeMethodInfo.Invoke (System.Object obj, System.Reflection.BindingFlags invokeAttr, System.Reflection.Binder binder, System.Object[] parameters, System.Globalization.CultureInfo culture) [0x0006a] in /Users/builder/jenkins/workspace/test-mono-pull-request-amd64-osx/mcs/class/corlib/System.Reflection/RuntimeMethodInfo.cs:395 
```