@filipnavara 2019-10-09
Preliminary performance numbers:

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_String.Substring_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 10, i2: |      8.99 |           107.84 |            12.00 |         |
| System.Tests.Perf_String.Substring_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 7, i2:  |      8.57 |           107.96 |            12.59 |         |
| System.Tests.Perf_String.Substring_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 0, i2:  |      8.43 |           109.22 |            12.96 |         |
| System.Tests.Perf_String.Remove_Int(s: "dzsdzsDDZSDZSDZSddsz", i: 7)             |      7.77 |           111.63 |            14.37 |         |
| System.Tests.Perf_String.Substring_Int(s: "dzsdzsDDZSDZSDZSddsz", i: 10)         |      7.39 |           114.61 |            15.51 |         |
| System.Tests.Perf_String.Substring_Int(s: "dzsdzsDDZSDZSDZSddsz", i: 7)          |      7.35 |           114.51 |            15.58 |         |
| System.Tests.Perf_String.Remove_Int(s: "dzsdzsDDZSDZSDZSddsz", i: 10)            |      7.23 |           113.58 |            15.71 |         |
| System.Tests.Perf_String.Remove_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 7, i2: 4)  |      6.43 |           114.17 |            17.76 |         |
| System.Tests.Perf_String.Remove_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 0, i2: 8)  |      6.36 |           112.68 |            17.72 |         |
| System.Tests.Perf_String.Remove_IntInt(s: "dzsdzsDDZSDZSDZSddsz", i1: 10, i2: 1) |      6.34 |           120.25 |            18.96 |         |
| System.Tests.Perf_String.Split(s: "A B C D E F G H I J K L M N O P Q R S T U V W |      6.09 |          3062.62 |           502.68 |         |
| System.Tests.Perf_String.TrimStart_CharArr(s: " Test", c: [' ', ' '])            |      6.02 |           116.02 |            19.27 |         |
| System.Tests.Perf_String.TrimEnd_CharArr(s: "Test ", c: [' ', ' '])              |      5.98 |           118.18 |            19.75 |         |
| System.Tests.Perf_String.Split(s: "A B C D E F G H I J K L M N O P Q R S T U V W |      5.88 |          3096.21 |           526.37 |         |
| System.Tests.Perf_String.Trim_CharArr(s: " Test", c: [' ', ' '])                 |      5.48 |           120.00 |            21.91 |         |
| System.Tests.Perf_String.Insert(s1: "Test", i: 2, s2: " Test")                   |      5.39 |           114.81 |            21.31 |         |
| System.Tests.Perf_String.Trim_CharArr(s: "Test ", c: [' ', ' '])                 |      5.33 |           121.05 |            22.73 |         |
| System.Tests.Perf_String.Insert(s1: "dzsdzsDDZSDZSDZSddsz", i: 7, s2: "Test")    |      5.11 |           117.97 |            23.08 |         |
| System.Tests.Perf_String.PadLeft(n: 18)                                          |      4.98 |           121.90 |            24.47 |         |
| PerfLabTests.LowLevelPerf.IntegerFormatting                                      |      4.79 |      12186752.20 |       2546510.19 |         |
| System.Tests.Perf_String.Trim_CharArr(s: " Te st  ", c: [' ', ' '])              |      4.63 |           121.91 |            26.31 |         |
| System.Tests.Perf_String.Replace_Char(text: "This is a very nice sentence", oldC |      3.36 |           139.86 |            41.60 |         |
| System.Tests.Perf_String.ToLowerInvariant(s: "TEST")                             |      3.23 |           144.91 |            44.86 |         |
| System.Tests.Perf_String.ToLowerInvariant(s: "TeSt")                             |      3.16 |           142.72 |            45.10 |         |
| System.Tests.Perf_String.ToUpperInvariant(s: "TeSt")                             |      3.16 |           140.87 |            44.63 |         |
| System.Tests.Perf_String.TrimEnd(s: "Test ")                                     |      3.15 |           145.54 |            46.21 |         |
| System.Tests.Perf_String.ToUpperInvariant(s: "test")                             |      3.10 |           138.51 |            44.67 |         |
| System.Tests.Perf_String.TrimStart(s: " Test")                                   |      3.09 |           143.10 |            46.29 |         |
| System.Tests.Perf_String.ToUpper(s: "TeSt")                                      |      2.98 |           141.85 |            47.60 |         |
| System.Tests.Perf_String.ToUpper(s: "test")                                      |      2.95 |           140.08 |            47.50 |         |
| System.Tests.Perf_String.Trim(s: " Test")                                        |      2.53 |           160.15 |            63.20 |         |
| System.Tests.Perf_String.Trim(s: "Test ")                                        |      2.53 |           160.01 |            63.28 |         |
| System.Tests.Perf_String.ToLower(s: "TEST")                                      |      2.28 |           143.54 |            62.93 |         |
| System.Tests.Perf_String.Concat_str_str(size: 100)                               |      2.25 |           172.50 |            76.50 |         |
| System.Tests.Perf_String.ToLower(s: "TeSt")                                      |      2.18 |           139.99 |            64.13 |         |
| System.Tests.Perf_String.Trim(s: " Te st  ")                                     |      2.01 |           194.77 |            96.70 |         |
| System.Tests.Perf_String.Concat_str_str_str(size: 100)                           |      1.82 |           206.83 |           113.63 |         |
| System.Tests.Perf_String.Replace_String(text: "This is a very nice sentence", ol |      1.77 |           216.99 |           122.28 |         |
| System.Tests.Perf_String.Format_MultipleArgs                                     |      1.64 |           511.59 |           311.07 |         |
| System.Tests.Perf_String.Interpolation_MultipleArgs(c: '1', s: "Foo")            |      1.64 |           512.00 |           312.67 |         |
| System.Tests.Perf_String.ToLowerInvariant(s: "This is a much longer piece of tex |      1.61 |           259.64 |           161.07 |         |
| System.Tests.Perf_String.ToUpperInvariant(s: "This is a much longer piece of tex |      1.61 |           253.11 |           157.64 |         |
| System.Tests.Perf_String.ToUpper(s: "This is a much longer piece of text that mi |      1.57 |           252.92 |           160.72 |         |
| System.Tests.Perf_String.ToLower(s: "This is a much longer piece of text that mi |      1.46 |           259.31 |           178.06 |         |
| System.Tests.Perf_String.Concat_str_str_str_str(size: 100)                       |      1.43 |           267.58 |           187.06 |         |